requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncwhois==1.1.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
fastapi==0.110.3
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
geoip2==4.8.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
maxminddb==2.6.1
motor==3.4.0
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycares==4.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.4.4
python-whois==0.9.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
threadpoolctl==3.5.0
tldextract==5.1.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whodap==0.1.12
yarl==1.9.4
